Mock JSON and refactor of types
