+++ src/__test__/mocks/image-info/1/image.json
@@ -1,7 +1,10 @@
 {
-  "entry_number": 1,
-  "pokemon_species": {
-    "name": "bulbasaur",
-    "url": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png"
-  }
+  "name": "image1",
+  "uploader": "Test_User",
+  "url": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png",
+  "tags": [
+    "HAS_SUN",
+    "HAS_ASTEROID",
+    "HAS_STAR"
+  ]
 }
