+++ src/api/fetchImageList.test.ts
@@ -4,7 +4,7 @@ import {
 } from "src/__test__/utils/suppress-unsuppress-errors";
 import fetchImageList from "src/api/fetchImageList";
 import { imageList500Factory } from "src/__test__/mocks/handlers";
-import images from "src/__test__/mocks/image-lists/images.json";
+import images from "src/__test__/mocks/image-lists/imagesPlaceholder.json";
 import { server } from "src/__test__/mocks/server";
 
 describe("fetchImageList", () => {
