+++ src/api/fetchImageList.ts
@@ -1,6 +1,6 @@
 import { challengeApi } from "../config";
 import headers from "src/api/headers";
-import { PokedexResult } from "src/types";
+import { PokedexResult } from "src/types/placeholder";
 
 const fetchImageList = async (): Promise<PokedexResult> => {
   const result = await fetch(`${challengeApi.baseUrl}/pokedex/1`, { headers });
