+++ src/api/headers.ts
@@ -1,5 +1,5 @@
 import { challengeApi } from "../config";
-import { Header } from "src/types";
+import { Header } from "src/types/header";
 
 const headers: Header = {
   "x-api-key": challengeApi.apiKey as string,
