+++ src/types/header.ts
@@ -1,0 +1,3 @@
+export type Header = {
+  "x-api-key": string;
+};
