+++ src/types/images.ts
@@ -1,0 +1,7 @@
+export type ImageResponse = {
+    id: string;
+    name: string;
+    imgURL: string;
+};
+
+export type ImageList = ImageResponse[];
